+++ CHANGELOG.md
@@ -39,6 +39,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 * Add missing `chfl_topology_resize` function to C API.
 * C API functions taking three lengths/angles now take a `double[3]` parameter
   instead.
+* Rename `chfl_topology_are_linked` to `chfl_topology_residues_linked`.
 
 ### Deprecation and removals